+++ Classes/Domain/Model/Filter/ConferenceFilter.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JS\Wechange\Domain\Model\Filter;
+
+class ConferenceFilter extends AbstractSortableFilter
+{
+}