+++ Classes/Domain/Model/Filter/FilterFactory.php
@@ -6,6 +6,9 @@ namespace JS\Wechange\Domain\Model\Filter;
 
 use JS\Wechange\Domain\Model\Coordinate;
 
+/**
+ * @codeCoverageIgnore
+ */
 class FilterFactory
 {
     public function makeMapFilter(array $coordinates, array $filterSettings): MapFilter
@@ -29,8 +32,46 @@ class FilterFactory
             (int)$filterSettings['parent'],
             $filterSettings['tag'],
             (int)$filterSettings['limit'],
+            (int)$filterSettings['offset'],
             $filterSettings['orderBy'],
             $filterSettings['orderDir']
         );
+    }
+
+    public function makeEventFilter(array $filterSettings): EventFilter
+    {
+        return new EventFilter(
+            (int)$filterSettings['limit'],
+            (int)$filterSettings['offset'],
+            $filterSettings['orderBy'],
+            $filterSettings['orderDir'],
+            (bool)$filterSettings['onlyUpcoming']
+        );
+    }
+
+    public function makeConferenceFilter(array $filterSettings): ConferenceFilter
+    {
+        return new ConferenceFilter(
+            (int)$filterSettings['limit'],
+            (int)$filterSettings['offset'],
+            $filterSettings['orderBy'],
+            $filterSettings['orderDir']
+        );
+    }
+
+    public function makeNoteFilter(array $filterSettings): NoteFilter
+    {
+        return new NoteFilter(
+            (int)$filterSettings['group'],
+            (int)$filterSettings['limit'],
+            (int)$filterSettings['offset'],
+            $filterSettings['orderBy'],
+            $filterSettings['orderDir']
+        );
+    }
+
+    public function makeStatisticFilter(): StatisticFilter
+    {
+        return new StatisticFilter();
     }
 }