+++ ext_emconf.php
@@ -4,7 +4,7 @@ $EM_CONF[$_EXTKEY] = [
     'title' => 'WECHANGE Integration',
     'description' => 'TYPO3 integration for the WECHANGE platform (https://wechange.de/)',
     'category' => 'plugin',
-    'version' => '1.1.0',
+    'version' => '2.0.0',
     'module' => '',
     'state' => 'stable',
     'uploadfolder' => '',
@@ -13,8 +13,7 @@ $EM_CONF[$_EXTKEY] = [
     'author_email' => 'mail@schlierenkamp.de',
     'constraints' => [
         'depends' => [
-            'typo3' => '9.5.9-10.4.99',
-            'php' => '7.4.0-7.4.99'
+            'typo3' => '10.4.0-11.5.99'
         ],
         'conflicts' => [
         ],