+++ ext_localconf.php
@@ -1,67 +1,70 @@
 <?php
+
 defined('TYPO3_MODE') || die();
 call_user_func(
     static function () {
         \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
-            'JS.Wechange',
+            'Wechange',
             'Projects',
             [
-                'Project' => 'list'
+                \JS\Wechange\Controller\ProjectController::class => 'list'
             ]
         );
 
         \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
-            'JS.Wechange',
-            'Map',
+            'Wechange',
+            'Events',
             [
-                'Map' => 'show'
+                \JS\Wechange\Controller\EventController::class => 'list'
             ]
         );
 
-        \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPageTSConfig(
-            'mod {
-                wizards.newContentElement.wizardItems.plugins {
-                    elements {
-                        projects {
-                            iconIdentifier = wechange-plugin-projects
-                            title = LLL:EXT:wechange/Resources/Private/Language/locallang_db.xlf:plugin.projects
-                            description = LLL:EXT:wechange/Resources/Private/Language/locallang_db.xlf:plugin.projects.description
-                            tt_content_defValues {
-                                CType = list
-                                list_type = wechange_projects
-                            }
-                        }
-                        map {
-                            iconIdentifier = wechange-plugin-map
-                            title = LLL:EXT:wechange/Resources/Private/Language/locallang_db.xlf:plugin.map
-                            description = LLL:EXT:wechange/Resources/Private/Language/locallang_db.xlf:plugin.map.description
-                            tt_content_defValues {
-                                CType = list
-                                list_type = wechange_map
-                            }
-                        }
-                    }
-                    show = *
-                }
-           }'
+        \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
+            'Wechange',
+            'Conferences',
+            [
+                \JS\Wechange\Controller\ConferenceController::class => 'list'
+            ]
+        );
+
+        \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
+            'Wechange',
+            'Notes',
+            [
+                \JS\Wechange\Controller\NoteController::class => 'list'
+            ]
+        );
+
+        \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
+            'Wechange',
+            'Statistics',
+            [
+                \JS\Wechange\Controller\StatisticController::class => 'list'
+            ]
+        );
+
+        \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
+            'Wechange',
+            'Map',
+            [
+                \JS\Wechange\Controller\MapController::class => 'show'
+            ]
         );
 
         /** @var \TYPO3\CMS\Core\Imaging\IconRegistry $iconRegistry */
         $iconRegistry = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(
             \TYPO3\CMS\Core\Imaging\IconRegistry::class
         );
+        $iconRegistry->registerIcon(
+            'wechange-plugin',
+            \TYPO3\CMS\Core\Imaging\IconProvider\BitmapIconProvider::class,
+            ['source' => 'EXT:wechange/Resources/Public/Icons/wechange.png']
+        );
 
-        foreach (['projects', 'map'] as $plugin) {
-            $iconRegistry->registerIcon(
-                'wechange-plugin-' . $plugin,
-                \TYPO3\CMS\Core\Imaging\IconProvider\BitmapIconProvider::class,
-                ['source' => 'EXT:wechange/Resources/Public/Icons/wechange.png']
-            );
-        }
+        \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPageTSConfig(
+            '<INCLUDE_TYPOSCRIPT: source="FILE:EXT:wechange/Configuration/TsConfig/All.tsconfig">'
+        );
 
-        if (!isset($GLOBALS['TYPO3_CONF_VARS']['SYS']['caching']['cacheConfigurations']['jswechange_projectList']) ||
-            !is_array($GLOBALS['TYPO3_CONF_VARS']['SYS']['caching']['cacheConfigurations']['jswechange_projectList'])) {
-            $GLOBALS['TYPO3_CONF_VARS']['SYS']['caching']['cacheConfigurations']['jswechange_projectList'] = [];
-        }
+        $GLOBALS['TYPO3_CONF_VARS']['SYS']['caching']['cacheConfigurations']['wechange_apirequests'] ??= [];
     }
 );